# contracts/api/openapi.yaml
...
components:
  schemas:
...
    EvidenceRelationMutationOut:
      properties:
        affected_thesis_ids:
          items:
            type: string
          title: Affected Thesis Ids
          type: array
        relation:
          $ref: '#/components/schemas/EvidenceRelationOut'
      required:
      - relation
      - affected_thesis_ids
      title: EvidenceRelationMutationOut
      type: object
    EvidenceRelationOut:
      description: 当前单关联模型的只读兼容输出，为多关联表迁移预留 response 形状。
      properties:
        can_manage:
          title: Can Manage
          type: boolean
        created_by:
          title: Created By
          type: string
        deactivated_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Deactivated At
        deactivated_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Deactivated By
        direction:
          title: Direction
          type: string
        hypothesis_id:
          title: Hypothesis Id
          type: string
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        relation_id:
          title: Relation Id
          type: string
        reviewed_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Reviewed At
        reviewed_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Reviewed By
        status:
          title: Status
          type: string
        strength:
          anyOf:
          - type: string
          - type: 'null'
          title: Strength
        thesis_id:
          title: Thesis Id
          type: string
      required:
      - relation_id
      - thesis_id
      - hypothesis_id
      - direction
      - status
      - created_by
      - can_manage
      title: EvidenceRelationOut
      type: object
    EvidenceRelationReviewIn:
      properties:
        action:
          pattern: ^(确认|驳回|暂不判断)$
          title: Action
          type: string
        reason:
          anyOf:
          - maxLength: 1000
            type: string
          - type: 'null'
          title: Reason
      required:
      - action
      title: EvidenceRelationReviewIn
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    IngestionReviewOut:
      properties:
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        document_id:
          title: Document Id
          type: string
        event_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Event Id
        job_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Job Id
        payload:
          title: Payload
          type: object
        reason:
          title: Reason
          type: string
        resolution:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolution
        resolved_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Resolved At
        review_id:
          title: Review Id
          type: string
        review_type:
          title: Review Type
          type: string
        security_candidates:
          items:
            type: object
          title: Security Candidates
          type: array
        status:
          title: Status
          type: string
      required:
      - review_id
      - review_type
      - document_id
      - job_id
      - event_id
      - reason
      - status
      - payload
      - security_candidates
      - resolution
      - created_at
      - resolved_at
      title: IngestionReviewOut
      type: object
    IngestionReviewResolveIn:
      properties:
        resolution:
          maxLength: 2000
          minLength: 2
          title: Resolution
          type: string
        security_id:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
          title: Security Id
      required:
      - resolution
      title: IngestionReviewResolveIn
      type: object
...
    ReviewTaskCreateIn:
      properties:
        assignee:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
          title: Assignee
        detail:
          anyOf:
          - type: object
          - type: 'null'
          title: Detail
        priority:
          default: 普通
          enum:
          - 低
          - 普通
          - 高
          - 紧急
          title: Priority
          type: string
        thesis_id:
          maxLength: 64
          minLength: 1
          title: Thesis Id
          type: string
        trigger:
          default: 人工发起
          enum:
          - 到期
          - 重大事件
          - 失效条件
          - 人工发起
          - 低置信
          - 处理失败
          title: Trigger
          type: string
      required:
      - thesis_id
      title: ReviewTaskCreateIn
      type: object
    ReviewTaskOut:
      properties:
        assignee:
          title: Assignee
          type: string
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        detail:
          anyOf:
          - type: object
          - type: 'null'
          title: Detail
        priority:
          title: Priority
          type: string
        resolution:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolution
        resolved_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Resolved At
        state:
          title: State
          type: string
        task_id:
          title: Task Id
          type: string
        thesis_id:
          title: Thesis Id
          type: string
        trigger:
          title: Trigger
          type: string
      required:
      - task_id
      - thesis_id
      - trigger
      - priority
      - assignee
      - state
      - detail
      - resolution
      - created_at
      - resolved_at
      title: ReviewTaskOut
      type: object
    ReviewTaskResolveIn:
      properties:
        resolution:
          maxLength: 2000
          minLength: 2
          title: Resolution
          type: string
      required:
      - resolution
      title: ReviewTaskResolveIn
      type: object
...
    SecurityIn:
      properties:
        aliases:
          items:
            type: string
          maxItems: 20
          title: Aliases
          type: array
        industry:
          anyOf:
          - maxLength: 128
            type: string
          - type: 'null'
          title: Industry
        name:
          maxLength: 255
          minLength: 1
          title: Name
          type: string
        security_id:
          maxLength: 64
          minLength: 1
          title: Security Id
          type: string
        ticker:
          anyOf:
          - maxLength: 32
            type: string
          - type: 'null'
          title: Ticker
      required:
      - security_id
      - name
      title: SecurityIn
      type: object
    SecurityOut:
      properties:
        aliases:
          items:
            type: string
          title: Aliases
          type: array
        industry:
          anyOf:
          - type: string
          - type: 'null'
          title: Industry
        name:
          title: Name
          type: string
        security_id:
          title: Security Id
          type: string
        ticker:
          anyOf:
          - type: string
          - type: 'null'
          title: Ticker
      required:
      - security_id
      - name
      title: SecurityOut
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  description: 主动权益投资逻辑智能协作平台。系统输出候选信号与状态建议，不产生任何交易、评级或调仓指令。
  title: AI Investment Copilot
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/evidence/{evidence_id}/relations/{relation_id}/review:
    post:
      operationId: review_relation_api_evidence__evidence_id__relations__relation_id__review_post
      parameters:
      - in: path
        name: evidence_id
        required: true
        schema:
          title: Evidence Id
          type: string
      - in: path
        name: relation_id
        required: true
        schema:
          title: Relation Id
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceRelationReviewIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceRelationMutationOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Review Relation
      tags:
      - thesis
...
  /api/reviews:
    get:
      operationId: list_reviews_api_reviews_get
      parameters:
      - in: query
        name: state
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: State
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ReviewTaskOut'
                title: Response List Reviews Api Reviews Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Reviews
      tags:
      - reviews
    post:
      operationId: create_review_api_reviews_post
      parameters:
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewTaskCreateIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTaskOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Review
      tags:
      - reviews
...
  /api/reviews/ingestion:
    get:
      operationId: list_ingestion_reviews_api_reviews_ingestion_get
      parameters:
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/IngestionReviewOut'
                title: Response List Ingestion Reviews Api Reviews Ingestion Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Ingestion Reviews
      tags:
      - reviews
  /api/reviews/ingestion/{review_id}/resolve:
    post:
      operationId: resolve_ingestion_review_api_reviews_ingestion__review_id__resolve_post
      parameters:
      - in: path
        name: review_id
        required: true
        schema:
          title: Review Id
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestionReviewResolveIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionReviewOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Resolve Ingestion Review
      tags:
      - reviews
  /api/reviews/{task_id}:
    get:
      operationId: get_review_api_reviews__task_id__get
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          title: Task Id
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTaskOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Review
      tags:
      - reviews
  /api/reviews/{task_id}/resolve:
    post:
      operationId: resolve_review_api_reviews__task_id__resolve_post
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          title: Task Id
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewTaskResolveIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTaskOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Resolve Review
      tags:
      - reviews
  /api/securities:
    get:
      operationId: list_securities_api_securities_get
      parameters:
      - in: query
        name: keyword
        required: false
        schema:
          anyOf:
          - maxLength: 100
            type: string
          - type: 'null'
          title: Keyword
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SecurityOut'
                title: Response List Securities Api Securities Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Securities
      tags:
      - securities
    post:
      operationId: create_security_api_securities_post
      parameters:
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - in: header
        name: X-User-Teams
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Teams
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Security
      tags:
      - securities
...